Extract a policy evaluation ingredient from the train ingredient.
