+++ benchmarking/example_airl_seals_swimmer_best_hp_eval.json
@@ -49,7 +49,6 @@
   },
   "total_timesteps": 10000000,
   "train": {
-    "n_episodes_eval": 50,
     "policy_cls": "MlpPolicy",
     "policy_kwargs": {
       "activation_fn": {
@@ -76,6 +75,9 @@
         }
       ]
     }
+  },
+  "policy_evaluation": {
+    "n_episodes_eval": 50
   },
   "environment": {
     "gym_id": "seals/Swimmer-v0"
