+++ benchmarking/example_bc_seals_ant_best_hp_eval.json
@@ -24,7 +24,6 @@
     "n_expert_demos": null
   },
   "train": {
-    "n_episodes_eval": 50,
     "policy_cls": {
       "py/type": "imitation.policies.base.FeedForward32Policy"
     },
@@ -38,6 +37,9 @@
         }
       }
     }
+  },
+  "policy_evaluation": {
+    "n_episodes_eval": 50
   },
   "environment": {
     "gym_id": "seals/Ant-v0"
