+++ benchmarking/example_gail_seals_ant_best_hp_eval.json
@@ -43,7 +43,6 @@
   },
   "total_timesteps": 10000000,
   "train": {
-    "n_episodes_eval": 50,
     "policy_cls": {
       "py/type": "imitation.policies.base.FeedForward32Policy"
     },
@@ -57,6 +56,9 @@
         }
       }
     }
+  },
+  "policy_evaluation": {
+    "n_episodes_eval": 50
   },
   "environment": {
     "gym_id": "seals/Ant-v0"
