+++ experiments/dagger_benchmark.sh
@@ -20,7 +20,7 @@ while true; do
     # Fast mode (debug)
     -f | --fast)
       SEEDS=(0)
-      extra_configs=("${extra_configs[@]}" environment.fast demonstrations.fast train.fast fast)
+      extra_configs=("${extra_configs[@]}" environment.fast demonstrations.fast policy_evaluation.fast fast)
       shift
       ;;
     --paper)  # Table benchmark settings
