+++ src/imitation/scripts/config/train_preference_comparisons.py
@@ -5,7 +5,7 @@ import sacred
 from imitation.algorithms import preference_comparisons
 from imitation.scripts.ingredients import environment
 from imitation.scripts.ingredients import logging as logging_ingredient
-from imitation.scripts.ingredients import reward, rl, train
+from imitation.scripts.ingredients import policy_evaluation, reward, rl, train
 
 train_preference_comparisons_ex = sacred.Experiment(
     "train_preference_comparisons",
@@ -15,6 +15,7 @@ train_preference_comparisons_ex = sacred.Experiment(
         reward.reward_ingredient,
         rl.rl_ingredient,
         train.train_ingredient,
+        policy_evaluation.policy_evaluation_ingredient,
     ],
 )
